// sahayak/sahayak/src/data/types.ts
export type Role =
  | 'citizen'
  | 'farmer'
  | 'health_worker'
  | 'student'
  | 'faculty'
  | 'industry'
  | 'admin'